vesuvius: promote neural_tracing disk chunk cache to data/chunk_cache
